refactoring getReport(city) into a lib for code sharing
