+++ src/weather-lib/metarfetcher.h
@@ -1,0 +1,9 @@
+#ifndef METARFETCHER_H
+#define METARFETCHER_H
+
+#include <QObject>
+
+QString getReport(QString city);
+
+
+#endif // METARFETCHER_H
