+++ src/weather-lib/weather-lib.pri
@@ -1,0 +1,3 @@
+INCLUDEPATH += $$IN_PWD
+LIBS += -L$$OUT_PWD/../weather-lib -lweather-lib
+
