+++ src/weather-lib/weather-lib.pro
@@ -1,0 +1,8 @@
+TEMPLATE = lib
+CONFIG += staticlib
+
+HEADERS += \
+    metarfetcher.h
+
+SOURCES += \
+    metarfetcher.cpp
