+++ src/weather-qt/mainwindow.cpp
@@ -1,9 +1,6 @@
 #include "mainwindow.h"
 #include "buttonconfig.h"
 #include "ui_mainwindow.h"
-#include "curl/curl.h"
-#include "curl/easy.h"
-#include <curl/curlver.h>
 #include "iostream"
 #include <stdlib.h>
 #include <QDebug>
@@ -14,8 +11,8 @@
 #include <qpainter.h>
 #include <QtPrintSupport/qprinter.h>
 #include <QtPrintSupport>
+#include "metarfetcher.h"
 using namespace std;
-long response_code;
 
 
 MainWindow::MainWindow(QWidget *parent) :
@@ -56,98 +53,15 @@ void MainWindow::on_lineEdit_returnPressed()
     on_pushButton_clicked();
 }
 
-struct MemoryStruct {
-  char *memory;
-  size_t size;
-};
-
-static size_t WriteMemoryCallback(void *contents, size_t size, size_t nmemb, void *userp)
-{
-  size_t realsize = size * nmemb;
-  struct MemoryStruct *mem = (struct MemoryStruct *)userp;
-
-  mem->memory = (char*)realloc(mem->memory, mem->size + realsize + 1);
-  if(mem->memory == NULL) {
-    /* out of memory! */
-    printf("not enough memory (realloc returned NULL)\n");
-    return 0;
-  }
-
-  memcpy(&(mem->memory[mem->size]), contents, realsize);
-  mem->size += realsize;
-  mem->memory[mem->size] = 0;
-
-  return realsize;
-}
-
 void MainWindow::printReport(QString city)
 {
-     if(city.length() == 0)
-     {
-         return;
-     }
-     CURL *curl_handle = nullptr;
-     CURLcode res;
-
-     struct MemoryStruct chunk;
-
-     chunk.memory = (char*)malloc(1);  /* will be grown as needed by the realloc above */
-     chunk.size = 0;    /* no data at this point */
-
-     curl_global_init(CURL_GLOBAL_ALL);
-
-     /* init the curl session */
-       curl_handle = curl_easy_init();
-     ostringstream os;
-     os << "http://tgftp.nws.noaa.gov/data/observations/metar/decoded/" << city.toStdString() << ".TXT";
-     string s =os.str();
-     cout << s;
-
-     /* specify URL to get */
-     curl_easy_setopt(curl_handle, CURLOPT_URL, s.c_str());
-
-     /* send all data to this function  */
-     curl_easy_setopt(curl_handle, CURLOPT_WRITEFUNCTION, WriteMemoryCallback);
-
-     /* we pass our 'chunk' struct to the callback function */
-     curl_easy_setopt(curl_handle, CURLOPT_WRITEDATA, (void *)&chunk);
-
-     /* some servers don't like requests that are made without a user-agent
-        field, so we provide one */
-     curl_easy_setopt(curl_handle, CURLOPT_USERAGENT, "libcurl-agent/1.0");
-
-     /* get it! */
-     res = curl_easy_perform(curl_handle);
-
-     /* check for errors */
-     if(res != CURLE_OK) {
-       fprintf(stderr, "curl_easy_perform() failed: %s\n",
-               curl_easy_strerror(res));
-     }
-
-
-     else {
-       /*
-        * Now, our chunk.memory points to a memory block that is chunk.size
-        * bytes big and contains the remote file.
-        *
-        * Do something nice with it!
-        */
-         curl_easy_getinfo(curl_handle, CURLINFO_RESPONSE_CODE, &response_code);
-         qDebug() << response_code;
-
-       printf("%lu bytes retrieved\n", (long)chunk.size);
-     }
-
-     /* cleanup curl stuff */
-     curl_easy_cleanup(curl_handle);
+    QString weather = getReport(city);
    //  curl_handle.
-     if(response_code == 200) {
+     if( weather != "") {
          statusBar()->showMessage("Printing report for " + city);
 
 
-         qDebug() << chunk.memory;
-         QString weather(chunk.memory);
+         qDebug() << weather;
 
          QPrinter printer;
          //printer.setOutputFileName("weather");
@@ -165,14 +79,6 @@ void MainWindow::printReport(QString city)
      else {
          statusBar()->showMessage("Invalid Airport code or other error");
      }
-
-
-     free(chunk.memory);
-
-     /* we're done with libcurl, so clean it up */
-     curl_global_cleanup();
-
-
 }
